+++ noodles-fasta/src/io/writer/builder.rs
@@ -32,9 +32,9 @@ impl Builder {
     /// ```
     /// # use std::io;
     /// use noodles_fasta as fasta;
-    /// let writer = fasta::io::writer::Builder::default().build_with_writer(io::sink());
+    /// let writer = fasta::io::writer::Builder::default().build_from_writer(io::sink());
     /// ```
-    pub fn build_with_writer<W>(self, writer: W) -> Writer<W>
+    pub fn build_from_writer<W>(self, writer: W) -> Writer<W>
     where
         W: Write,
     {
@@ -42,6 +42,15 @@ impl Builder {
             inner: writer,
             line_base_count: self.line_base_count,
         }
+    }
+
+    /// Builds a FASTA writer from a writer.
+    #[deprecated(since = "0.43.0", note = "Use `Builder::build_from_writer` instead.")]
+    pub fn build_with_writer<W>(self, writer: W) -> Writer<W>
+    where
+        W: Write,
+    {
+        self.build_from_writer(writer)
     }
 }